+++ Housing.Core/Entities/ViewModels/StudentViewModel.cs
@@ -11,7 +11,7 @@ namespace UWHousing.Entities.ViewModels
     /// </summary>
     public class StudentViewModel
     {
-        public int StudentID { get; set; }
+        public long StudentID { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string BuildingName { get; set; }
